[Upd] Refactor getting rows and rowStatements to the TabularReader from TabularModule
