+++ s-pipes-modules/module-tabular/src/main/java/cz/cvut/spipes/modules/tabular/TabularReader.java
@@ -4,14 +4,15 @@ import cz.cvut.spipes.modules.Mode;
 import cz.cvut.spipes.modules.model.Column;
 import cz.cvut.spipes.modules.model.Row;
 import cz.cvut.spipes.modules.model.TableSchema;
+import org.apache.jena.rdf.model.Statement;
 
 import java.io.IOException;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Set;
 
 public interface TabularReader {
     List<String> getHeader() throws IOException;
-    ArrayList<Column> getOutputColumns(List<String>header);
-    Set<Row> getRows(TableSchema tableSchema, String sourceResourceUri, Mode outputMode) throws IOException;
+    List<Column> getOutputColumns(List<String>header);
+    Set<Row> getRows(TableSchema tableSchema, String sourceResourceUri, Mode outputMode);
+    List<Statement> getRowStatements(List<String>header, List<Column>outputColumns, TableSchema tableSchema) throws IOException;
 }
